factored update_in out of Array and Hash into Associable module; consolidated specs; updated doc
