+++ lib/hamster.rb
@@ -9,3 +9,4 @@ require "hamster/sorted_set"
 require "hamster/mutable_hash"
 require "hamster/nested"
 require "hamster/version"
+require "hamster/associable"
